formatter: page_width: 100 remove button classes from view separate new_wallet data classes from view model extractract nested widgets into methods remove logic from ui
