+++ lib/view_model/security_backup_view_model.dart
@@ -46,6 +46,4 @@ class SecurityBackupViewModel extends ViewModel {
   ];
 
   CoinWallet wallet;
-
-  void titleUpdate(final String? suggestedTitle) {}
 }
