+++ lib/views/wallet_edit.dart
@@ -1,7 +1,7 @@
 import 'package:cupcake/coins/abstract/wallet_info.dart';
 import 'package:cupcake/view_model/wallet_edit_view_model.dart';
 import 'package:cupcake/views/abstract.dart';
-import 'package:cupcake/views/initial_setup_screen.dart';
+import 'package:cupcake/views/widgets/buttons/long_primary.dart';
 import 'package:cupcake/views/widgets/form_builder.dart';
 import 'package:flutter/material.dart';
 
@@ -31,21 +31,14 @@ class WalletEdit extends AbstractView {
               child: LongPrimaryButton(
                 backgroundColor: const WidgetStatePropertyAll(Colors.red),
                 icon: null,
-                onPressed: () async {
-                  await viewModel.deleteWallet();
-                },
+                onPressed: viewModel.deleteWallet,
                 text: "Delete",
               ),
             ),
             Expanded(
               child: LongPrimaryButton(
                 icon: null,
-                onPressed: () async {
-                  for (final element in viewModel.form) {
-                    if (!element.isOk) continue;
-                  }
-                  await viewModel.renameWallet();
-                },
+                onPressed: () => viewModel.renameWallet(),
                 text: "Rename",
               ),
             ),
